# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Registro de Pontos",
        "description": "API para gerenciar pontos dos funcionários",
        "contact": {
            "email": "diegohsananias@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/employees/{id}": {
            "put": {
                "tags": [
                    "Employees"
                ],
                "summary": "Atualizar um funcionário",
                "description": "Atualiza os dados de um funcionário existente com base no ID.",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Carlos Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "maxLength": 14,
                                        "example": "123.456.789-09"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.silva@example.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "employee"
                                        ],
                                        "example": "employee"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-20"
                                    },
                                    "address": {
                                        "properties": {
                                            "cep": {
                                                "type": "string",
                                                "maxLength": 9,
                                                "example": "01001-000"
                                            },
                                            "street": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Rua Exemplo"
                                            },
                                            "number": {
                                                "type": "integer",
                                                "example": 123
                                            },
                                            "city": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Cidade Exemplo"
                                            },
                                            "state": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Estado Exemplo"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Funcionário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Funcionário atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees"
                ],
                "summary": "Remover um funcionário",
                "description": "Remove um funcionário do sistema com base no ID.",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Funcionário removido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Funcionário removido com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                }
            }
        },
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Listar funcionários",
                "description": "Lista os funcionários cadastrados no sistema com filtros opcionais.",
                "operationId": "listEmployees",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nome do funcionário",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Carlos Silva"
                        }
                    },
                    {
                        "name": "cpf",
                        "in": "query",
                        "description": "Filtrar por CPF do funcionário",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "123.456.789-09"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Filtrar por e-mail do funcionário",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "example": "carlos.silva@example.com"
                        }
                    },
                    {
                        "name": "birth_date",
                        "in": "query",
                        "description": "Filtrar por data de nascimento do funcionário",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "1990-05-20"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de funcionários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Carlos Silva"
                                            },
                                            "cpf": {
                                                "type": "string",
                                                "example": "123.456.789-09"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "carlos.silva@example.com"
                                            },
                                            "role": {
                                                "type": "string",
                                                "example": "employee"
                                            },
                                            "birth_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "1990-05-20"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Cadastrar um funcionário",
                "description": "Cadastra um novo funcionário no sistema.",
                "operationId": "createEmployee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Carlos Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "maxLength": 14,
                                        "example": "123.456.789-09"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.silva@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "employee"
                                        ],
                                        "example": "employee"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-20"
                                    },
                                    "address": {
                                        "properties": {
                                            "cep": {
                                                "type": "string",
                                                "maxLength": 9,
                                                "example": "01001-000"
                                            },
                                            "street": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Rua Exemplo"
                                            },
                                            "number": {
                                                "type": "integer",
                                                "example": 123
                                            },
                                            "city": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Cidade Exemplo"
                                            },
                                            "state": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Estado Exemplo"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Funcionário cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Funcionário cadastrado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "409": {
                        "description": "CPF ou e-mail já está em uso"
                    }
                }
            }
        },
...
    },
...
}